backend/requirements.txt
pymongo
passlib[bcrypt]
python-jose
fastapi
uvicorn
python-multipart
langchain-openai
langgraph
python-dotenv
typing-extensions 